+++ src/backend_cairo/filter.rs
@@ -70,7 +70,7 @@ impl ImageExt for cairo::ImageSurface {
 
     fn into_srgb(&mut self) {
         if let Ok(ref mut data) = self.get_data() {
-            from_premultiplied(data);
+            filter::from_premultiplied(data.as_bgra_mut());
 
             for p in data.as_bgra_mut() {
                 p.r = filter::LINEAR_RGB_TO_SRGB_TABLE[p.r as usize];
@@ -78,7 +78,7 @@ impl ImageExt for cairo::ImageSurface {
                 p.b = filter::LINEAR_RGB_TO_SRGB_TABLE[p.b as usize];
             }
 
-            into_premultiplied(data);
+            filter::into_premultiplied(data.as_bgra_mut());
         } else {
             warn!("Cairo surface is already borrowed.");
         }
@@ -86,7 +86,7 @@ impl ImageExt for cairo::ImageSurface {
 
     fn into_linear_rgb(&mut self) {
         if let Ok(ref mut data) = self.get_data() {
-            from_premultiplied(data);
+            filter::from_premultiplied(data.as_bgra_mut());
 
             for p in data.as_bgra_mut() {
                 p.r = filter::SRGB_TO_LINEAR_RGB_TABLE[p.r as usize];
@@ -94,7 +94,7 @@ impl ImageExt for cairo::ImageSurface {
                 p.b = filter::SRGB_TO_LINEAR_RGB_TABLE[p.b as usize];
             }
 
-            into_premultiplied(data);
+            filter::into_premultiplied(data.as_bgra_mut());
         } else {
             warn!("Cairo surface is already borrowed.");
         }
@@ -120,28 +120,6 @@ fn copy_image(
     cr.paint();
 
     Ok(new_image)
-}
-
-pub(crate) fn from_premultiplied(data: &mut [u8]) {
-    // https://www.cairographics.org/manual/cairo-Image-Surfaces.html#cairo-format-t
-
-    for p in data.as_bgra_mut() {
-        let a = p.a as f64 / 255.0;
-        p.b = (p.b as f64 / a + 0.5) as u8;
-        p.g = (p.g as f64 / a + 0.5) as u8;
-        p.r = (p.r as f64 / a + 0.5) as u8;
-    }
-}
-
-fn into_premultiplied(data: &mut [u8]) {
-    // https://www.cairographics.org/manual/cairo-Image-Surfaces.html#cairo-format-t
-
-    for p in data.as_bgra_mut() {
-        let a = p.a as f64 / 255.0;
-        p.b = (p.b as f64 * a + 0.5) as u8;
-        p.g = (p.g as f64 * a + 0.5) as u8;
-        p.r = (p.r as f64 * a + 0.5) as u8;
-    }
 }
 
 struct CairoFilter;
@@ -207,7 +185,8 @@ impl Filter<cairo::ImageSurface> for CairoFilter {
         ts: &usvg::Transform,
         input: Image,
     ) -> Result<Image, Error> {
-        let (std_dx, std_dy) = try_opt_or!(Self::resolve_std_dev(fe, units, bbox, ts), Ok(input));
+        let (std_dx, std_dy, box_blur)
+            = try_opt_or!(Self::resolve_std_dev(fe, units, bbox, ts), Ok(input));
 
         let input = input.into_color_space(cs)?;
         let mut buffer = input.take()?;
@@ -215,9 +194,11 @@ impl Filter<cairo::ImageSurface> for CairoFilter {
         let (w, h) = (buffer.width(), buffer.height());
 
         if let Ok(ref mut data) = buffer.get_data() {
-            from_premultiplied(data);
-            filter::blur::apply(data, w, h, std_dx, std_dy, 4);
-            into_premultiplied(data);
+            if box_blur {
+                filter::box_blur::apply(data, w, h, std_dx, std_dy);
+            } else {
+                filter::iir_blur::apply(data, w, h, std_dx, std_dy);
+            }
         }
 
         Ok(Image::from_image(buffer, cs))
@@ -448,7 +429,7 @@ impl Filter<cairo::ImageSurface> for CairoFilter {
         let mut buffer = input.take()?;
 
         if let Ok(ref mut data) = buffer.get_data() {
-            from_premultiplied(data);
+            filter::from_premultiplied(data.as_bgra_mut());
 
             for pixel in data.as_bgra_mut() {
                 pixel.r = fe.func_r.apply(pixel.r);
@@ -457,7 +438,7 @@ impl Filter<cairo::ImageSurface> for CairoFilter {
                 pixel.a = fe.func_a.apply(pixel.a);
             }
 
-            into_premultiplied(data);
+            filter::into_premultiplied(data.as_bgra_mut());
         }
 
         Ok(Image::from_image(buffer, cs))
@@ -472,9 +453,9 @@ impl Filter<cairo::ImageSurface> for CairoFilter {
         let mut buffer = input.take()?;
 
         if let Ok(ref mut data) = buffer.get_data() {
-            from_premultiplied(data);
+            filter::from_premultiplied(data.as_bgra_mut());
             filter::color_matrix::apply(&fe.kind, data.as_bgra_mut());
-            into_premultiplied(data);
+            filter::into_premultiplied(data.as_bgra_mut());
         }
 
         Ok(Image::from_image(buffer, cs))